Add files via upload
The code of this Game generates a random number and then ask user to guess that random number. It also advises user to guess either high or low depending on user's last guess. In the end it provides user the count of attempts he/she took to guess the correct number